backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Vector DB
chromadb==0.5.5

# Google AI
google-generativeai==0.8.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.9

# Document parsing
python-docx==1.1.2

# HTTP
httpx==0.27.0
sse-starlette==2.1.0

# Utilities
python-dotenv==1.0.1
